// types/footerInfo.d.ts
export type TGQLFooterPhone = {
  id: string;
  type: string;
  phone: string;
};
export type TGQLFooterAddress = {
  id: string;
  type: string;
  address: string;
};
export type TGQLFooterMessage = {
  id: string;
  type: string;
  message: string;
};

export type TGQLFooterSocialMedia = {
  id: string;
  type: string;
  name: String;
  url: String;
  isEnabled: Boolean;
};